+++ lib/extreme/application.ex
@@ -1,21 +1,44 @@
 defmodule Commanded.EventStore.Adapters.Extreme.Application do
   use Application
 
-  alias Commanded.EventStore.Adapters.Extreme.Config
+  alias Commanded.EventStore.Adapters.Extreme.{
+    Config,
+    EventPublisher,
+    SubscriptionsSupervisor,
+    PubSub
+  }
+
+  @event_store Commanded.EventStore.Adapters.Extreme.EventStore
 
   def start(_type, _args) do
-    import Supervisor.Spec, warn: false
-
     children = [
-      worker(Extreme, [
-        Config.event_store_settings(),
-        [name: Commanded.EventStore.Adapters.Extreme.EventStore]
-      ]),
-      supervisor(Commanded.EventStore.Adapters.Extreme.SubscriptionsSupervisor, [])
+      {Registry, keys: :duplicate, name: PubSub, partitions: 1},
+      %{
+        id: Extreme,
+        start: {Extreme, :start_link, [Config.event_store_settings(), [name: @event_store]]},
+        restart: :permanent,
+        shutdown: 5000,
+        type: :worker
+      },
+      %{
+        id: EventPublisher,
+        start:
+          {EventPublisher, :start_link, [
+            @event_store,
+            "$ce-" <> Config.stream_prefix(),
+            [name: EventPublisher]
+          ]},
+        restart: :permanent,
+        shutdown: 5000,
+        type: :worker
+      },
+      SubscriptionsSupervisor
     ]
 
-    opts = [strategy: :one_for_one, name: Commanded.EventStore.Adapters.Extreme.Supervisor]
-
-    Supervisor.start_link(children, opts)
+    Supervisor.start_link(
+      children,
+      strategy: :one_for_one,
+      name: Commanded.EventStore.Adapters.Extreme.Supervisor
+    )
   end
 end